requirements.txt
asgiref==3.8.1
babel==2.17.0
bleach==6.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-money==3.5.3
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
filelock==3.15.4
fsspec==2024.6.1
huggingface-hub==0.24.5
idna==3.7
joblib==1.3.2
kaggle==1.6.14
numpy==1.26.1
packaging==24.1
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.11
py-moneyed==3.0
PyJWT==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2023.3.post1
PyYAML==6.0.1
reportlab==4.4.2
requests==2.32.3
scikit-learn==1.3.2
scipy==1.11.3
setuptools==78.1.0
six==1.16.0
sqlparse==0.5.3
text-unidecode==1.3
threadpoolctl==3.2.0
tqdm==4.66.4
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.2.1
webencodings==0.5.1
